+++ src/ts/modules/FastScreen.ts
@@ -8,7 +8,7 @@ class FastScreen {
   constructor() {
     this.create()
 
-    window.addEventListener(EVT.events.pageTypeChange, () => {
+    window.addEventListener(EVT.events.pageSwitchedTypeChange, () => {
       this.destroy()
     })
   }