+++ src/ts/modules/Log.ts
@@ -9,7 +9,7 @@ class Log {
     this.scrollToBottom()
 
     // 切换不同页面时，如果任务已经完成，则清空输出区域，避免日志一直堆积。
-    window.addEventListener(EVT.events.pageTypeChange, () => {
+    window.addEventListener(EVT.events.pageSwitchedTypeChange, () => {
       if (store.states.allowWork) {
         this.clear()
       }